Move release logic to the Releases tab
