+++ lib/nerves_hub_web/live/deployment_groups/show.ex
@@ -12,7 +12,7 @@ defmodule NervesHubWeb.Live.DeploymentGroups.Show do
   alias Phoenix.Socket.Broadcast
 
   alias NervesHubWeb.Components.DeploymentGroupPage.Activity, as: ActivityTab
-  alias NervesHubWeb.Components.DeploymentGroupPage.ReleaseHistory, as: ReleaseHistoryTab
+  alias NervesHubWeb.Components.DeploymentGroupPage.Releases, as: ReleasesTab
   alias NervesHubWeb.Components.DeploymentGroupPage.Settings, as: SettingsTab
   alias NervesHubWeb.Components.DeploymentGroupPage.Summary, as: SummaryTab
 
@@ -295,6 +295,13 @@ defmodule NervesHubWeb.Live.DeploymentGroups.Show do
     send_update(SummaryTab, id: "deployment_group_summary", delta_updated: true)
 
     {:noreply, socket}
+  end
+
+  @impl Phoenix.LiveView
+  def handle_info({:flash, level, message}, socket) do
+    socket
+    |> put_flash(level, message)
+    |> noreply()
   end
 
   defp selected_tab(socket) do
