+++ lib/nerves_hub_web/live/deployment_groups/show.html.heex
@@ -70,8 +70,8 @@
       <.link class={tab_classes(@tab, :summary)} patch={~p"/org/#{@org}/#{@product}/deployment_groups/#{@deployment_group}"}>
         Summary
       </.link>
-      <.link class={tab_classes(@tab, :release_history)} patch={~p"/org/#{@org}/#{@product}/deployment_groups/#{@deployment_group}/releases"}>
-        Release History
+      <.link class={tab_classes(@tab, :releases)} patch={~p"/org/#{@org}/#{@product}/deployment_groups/#{@deployment_group}/releases"}>
+        Releases
       </.link>
       <.link class={tab_classes(@tab, :activity)} patch={~p"/org/#{@org}/#{@product}/deployment_groups/#{@deployment_group}/activity"}>
         Activity
@@ -101,7 +101,17 @@
     update_stats={@update_stats}
   />
 
-  <.live_component :if={@tab == :release_history} module={ReleaseHistoryTab} id="deployment_group_release_history" releases={@releases} />
+  <.live_component
+    :if={@tab == :releases}
+    module={ReleasesTab}
+    id="deployment_group_releases"
+    releases={@releases}
+    deployment_group={@deployment_group}
+    org={@org}
+    product={@product}
+    user={@user}
+    org_user={@org_user}
+  />
 
   <.live_component :if={@tab == :activity} module={ActivityTab} id="deployment_group_activity" deployment_group={@deployment_group} org={@org} product={@product} user={@user} />
