+++ lib/nerves_hub_web/router.ex
@@ -373,7 +373,7 @@ defmodule NervesHubWeb.Router do
       live(
         "/org/:org_name/:product_name/deployment_groups/:name/releases",
         Live.DeploymentGroups.Show,
-        :release_history
+        :releases
       )
 
       live(
